add resize image functionality

+++ fields/types/gcsimage/GcsImageType.js
@@ -3,8 +3,9 @@
  */
 
 var _ = require('underscore');
-var grappling = require('grappling-hook');
+var fs = require('fs');
 var gcsHelper = require('../../../lib/gcsHelper');
+var grappling = require('grappling-hook');
 var keystone = require('../../../');
 var moment = require('moment');
 var super_ = require('../Type');
@@ -22,7 +23,7 @@ function extractExif(image) {
                     console.log('Error while extracting exif of image: ' + error.message);
                     resolve({});
                 } else {
-                    if ( typeof exifData === 'object') {
+                    if (typeof exifData === 'object') {
                         return resolve(exifData);
                     }
                     resolve({});
@@ -174,14 +175,20 @@ gcsimage.prototype.addToSchema = function() {
             var promise = new Promise(function(resolve, reject) {
                 var gcsConfig = field.gcsConfig;
                 var bucket = gcsHelper.initBucket(gcsConfig, _this.get(paths.bucket));
-                bucket.deleteFiles({
-                    prefix: _this.get(paths.path) + _this.get(paths.filename)
-                }, function(err) {
-                    if (err) {
-                        return reject(err);
-                    }
+                var filename = _this.get(paths.filename);
+                if (filename && typeof filename === 'string') {
+                    var filenameWithoutExt = filename.split('.');
+                    bucket.deleteFiles({
+                        prefix: _this.get(paths.path) + filenameWithoutExt
+                    }, function(err) {
+                        if (err) {
+                            return reject(err);
+                        }
+                        resolve();
+                    });
+                } else {
                     resolve();
-                });
+                }
             });
             reset(this);
             return promise;
@@ -273,6 +280,9 @@ gcsimage.prototype.uploadFile = function(item, file, update, callback) {
     var isPublicRead = field.options.publicRead ? field.options.publicRead : false;
     var prefix = field.options.datePrefix ? moment().format(field.options.datePrefix) + '-' : '';
     var filename = prefix + file.name;
+    var split = filename.split('.');
+    var filenameWithoutExt = split[0];
+    var ext = split[1] || '';
     var originalname = file.originalname;
     var filetype = file.mimetype || file.type;
 
@@ -293,11 +303,35 @@ gcsimage.prototype.uploadFile = function(item, file, update, callback) {
         }
 
         var bucket = gcsHelper.initBucket(field.gcsConfig, field.options.bucket);
-        gcsHelper.uploadFileToBucket(bucket, file.path, {
+        gcsHelper.uploadFileToBucket(bucket, fs.createReadStream(file.path), {
             destination: path + filename,
-        }).then(function(uploadedFile) {
-            return gcsHelper.makeFilePublicPrivateRead(uploadedFile, isPublicRead);
-        }).then(function(response) {
+            filetype: filetype,
+            isPublicRead: isPublicRead
+        }).then(function(apiResponse) {
+            // resizing image and upload resized images
+            if (typeof field.options.resize === 'function') {
+                var resizeFunc = field.options.resize;
+                if (Array.isArray(field.options.resizeOpts) && field.options.resizeOpts.length > 0) {
+                    var promises = [];
+                    var resizeOpts = field.options.resizeOpts;
+                    for (var i = 0; i < resizeOpts.length; i++) {
+                        var resizeOpt = resizeOpts[i] || {};
+                        promises.push(gcsHelper.uploadFileToBucket(bucket, resizeFunc(file.path, resizeOpt.width, resizeOpt.height, resizeOpt.options), {
+                            destination: path + filenameWithoutExt + '-' + resizeOpt.target + '.' + ext,
+                            filetype: filetype,
+                            isPublicRead: isPublicRead
+                        }));
+                    }
+                    return Promise.all(promises);
+                } else {
+                    // skip resizing
+                    return;
+                }
+            } else {
+                // skip resizing
+                return;
+            }
+        }).then(function(value) {
             extractExif(file.path).then(function(exifData) {
                 exifData.image = exifData.image || {};
                 exifData.exif = exifData.exif || {};
@@ -320,7 +354,14 @@ gcsimage.prototype.uploadFile = function(item, file, update, callback) {
                 return callback(null, fileData);
             });
         }).catch(function(err) {
-            return callback(err);
+            bucket.deleteFiles({
+                prefix: path + filenameWithoutExt
+            }, function(deleteErr) {
+                if (err) {
+                    return callback(deleteErr);
+                }
+                callback(err);
+            });
         });
     };
